Modify PATH directly rather than running settings64.sh
Docker uses sh as its login shell. If I source settings64.sh in .profile
for sh, I get errors, as `source` is used by the script and it is not
available to sh. So instead, set the PATH directly, rather than relying
on the script to do it for you.